+++ include/interrupt.h
@@ -1,16 +1,16 @@
-// © 2021 Qualcomm Innovation Center, Inc. All rights reserved.
+// Copyright © Qualcomm Technologies, Inc. and/or its subsidiaries.
 //
 // SPDX-License-Identifier: BSD-3-Clause
 
 #define IRQ_MAGIC 0x49U
 
-#define IOCTL_ENABLE_IRQ      IOW(IRQ_MAGIC, 0U, int)
-#define IOCTL_DISABLE_IRQ     IOW(IRQ_MAGIC, 1U, int)
+#define IOCTL_ENABLE_IRQ      IOW(IRQ_MAGIC, 0U, int32_t)
+#define IOCTL_DISABLE_IRQ     IOW(IRQ_MAGIC, 1U, int32_t)
 #define IOCTL_REGISTER_ISR    IOW(IRQ_MAGIC, 2U, struct register_isr_req)
 #define IOCTL_SET_IRQ_TRIGGER IOW(IRQ_MAGIC, 3U, struct irq_set_trigger_req)
-#define IOCTL_DEREGISTER_ISR  IOW(IRQ_MAGIC, 4U, int)
-#define IOCTL_ASSERT_IRQ      IOW(IRQ_MAGIC, 5U, int)
-#define IOCTL_CLEAR_IRQ	      IOW(IRQ_MAGIC, 6U, int)
+#define IOCTL_DEREGISTER_ISR  IOW(IRQ_MAGIC, 4U, int32_t)
+#define IOCTL_ASSERT_IRQ      IOW(IRQ_MAGIC, 5U, int32_t)
+#define IOCTL_CLEAR_IRQ	      IOW(IRQ_MAGIC, 6U, int32_t)
 
 typedef bool (*isr_t)(virq_t irq, void *data);
 
@@ -36,17 +36,17 @@ struct irq_set_trigger_req {
 void
 interrupt_init(void);
 
-int
+int32_t
 interrupt_register_isr(virq_t irq, isr_t isr, void *data);
 
-int
+int32_t
 interrupt_deregister_isr(virq_t irq);
 
 void
 interrupt_dispatch(void);
 
-long
-interrupt_open(int flags);
+int32_t
+interrupt_open(uint32_t flags);
 
 bool
 interrupt_wait(void);
